refactor: little cleanup before adding the new api
