+++ apps/router/src/types/index.tsx
@@ -5,3 +5,6 @@ export interface ServiceConfig {
     baseUrl: string;
   };
 }
+
+export const UNIT_OPTIONS = ['msats', 'sats', 'btc'] as const;
+export type Unit = (typeof UNIT_OPTIONS)[number];
